stress_mapper additions: I added in OESVM1 tables for cbeam, cbar, chexa, cpenta, ctria3, cquad4 and cbush. I imagine more can be added in the future (these were the only ones pertinent to the file I was trying to open).
Bug fixes: There were some bugs in the reading in of the cquad4 and ctria3 OESVM1 tables which had to do with the assumed number of words. I'm not sure if this is an MSC vs NX natran issue or the difference between real/imaginary and amplitude/phase complex data.